complete project

+++ Assets/Scripts/Health/Health.cs
@@ -29,7 +29,7 @@ public class Health : MonoBehaviour
         anim = GetComponent<Animator>();
         spriteRend = GetComponent<SpriteRenderer>();
     }
-    public void TakeDamage(float _damage)
+    public void TakeDamage(int _damage)
     {
         if (Invunerable) return;
         currentHealth = Mathf.Clamp(currentHealth - _damage, 0, startingHealth);
