+++ Assets/Scripts/Traps/EnemyDamage.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class EnemyDamage : MonoBehaviour
 {
-    [SerializeField] protected float damage;
+    [SerializeField] protected int damage;
 
     protected void OnTriggerEnter2D(Collider2D collision)
     {
